Remove tests for V8 internals
These test files are dedicated to V8 internal implementation details (or
regressions therein). Since they are not suitable for the Test262
project, they may be removed from this repository.

Note that assertions for implementation details may exist in files not
removed in this commit; this changeset is a preliminary coarse-grained
attempt to remove assertions not necessary for the Test262 migration.